+++ dbt/aus_personas/macros/abs/parse_abs_axis_helpers.sql
@@ -31,9 +31,11 @@ nullif(
         case
             when {{ logical_code_expr }} in ('G54') then regexp_replace({{ category_expr }}, '^Industry:\s*', '', 'i')
             when {{ logical_code_expr }} in ('G60') then regexp_replace({{ category_expr }}, '^Occupation:\s*', '', 'i')
+            when {{ logical_code_expr }} in ('G08') then regexp_replace({{ category_expr }}, '^Ancestry:\s*', '', 'i')
             when {{ logical_code_expr }} in ('G50') then regexp_replace({{ category_expr }}, '^Field of study:\s*', '', 'i')
             when {{ logical_code_expr }} in ('G36') then regexp_replace({{ category_expr }}, '^Dwelling structure:\s*', '', 'i')
             when {{ logical_code_expr }} in ('G17') then regexp_replace(regexp_replace({{ category_expr }}, '^Personal income:\s*', '', 'i'), '^Income:\s*', '', 'i')
+            when {{ logical_code_expr }} in ('G13') then regexp_replace({{ category_expr }}, '^Language:\s*', '', 'i')
             else {{ category_expr }}
         end
     ),